uploads/requirements.txt
# requirements.txt for Python 3.10

flask==2.3.3
tensorflow==2.15.0
keras==3.1.0
numpy==1.23.5
pillow==9.5.0